Ultima Entrega JS

+++ script/main.js
@@ -1,13 +1,41 @@
-
-// Verifica el nombre
+// Verifica si ya existe un nombre de usuario almacenado en localStorage
 let storedUsername = localStorage.getItem("user");
 
-// Si no hay un nombre de usuario almacenado, solicita uno
+// Si no hay un nombre de usuario almacenado, solicita uno usando Swal.fire
 if (!storedUsername) {
-    let newUsername = prompt("Ingrese Su Nombre");
-    localStorage.setItem("user", newUsername);
-    storedUsername = newUsername;
+  Swal.fire({
+    title: 'Ingrese su nombre',
+    input: 'text',
+    inputAttributes: {
+      autocapitalize: 'off'
+    },
+    confirmButtonText: 'Guardar',
+    showLoaderOnConfirm: true,
+    preConfirm: (newUsername) => {
+      // Valida si el nombre de usuario es válido
+      if (!newUsername) {
+        Swal.showValidationMessage('El nombre no puede estar en blanco');
+      } else {
+        // Almacena el nombre de usuario en localStorage
+        localStorage.setItem("user", newUsername);
+        return newUsername;
+      }
+    },
+    allowOutsideClick: () => !Swal.isLoading()
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire('Nombre guardado con éxito');
+      storedUsername = result.value; // Actualiza la variable almacenada con el nuevo nombre
+      
+      //recarga la pagina despues de recibir la respuesta , funcion asincrona (si se lo quito no me reiniciaba la pagina )
+      setTimeout(() => {
+        location.reload();
+      }, 2000);
+    }
+  });
 }
+
+// Ahora, puedes usar la variable storedUsername en tu código.
 
 // Selecciona el elemento para mostrar el nombre en el DOM
 let muestraNombre = document.getElementById("Muestranombre");
@@ -44,4 +72,5 @@ toggle.addEventListener("change", ()=>{
         localStorage.setItem("darkmode", false)
     }
 });
+
